requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
distro==1.9.0
Flask==3.1.3
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.120.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
google-genai==1.66.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.62.3
gspread==6.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
pillow==10.3.0
proto-plus==1.27.1
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
StrEnum==0.4.15
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.6
tenacity>=8.2.3

google-genai>=0.1.0
tenacity>=8.2.3
